use Popover instead of Modal
